refactor(example): extract TS/TSX tokenizer to shared lib
